Move IssuerInfo to pkg/shared
