+++ pkg/shared/issuerinfo.go
@@ -4,9 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package utils
-
-import "github.com/gardener/cert-management/pkg/shared"
+package shared
 
 const (
 	// IssuerTypeACME is the issuer type ACME
@@ -19,27 +17,27 @@ const (
 
 // IssuerInfo provides name and type of an issuer
 type IssuerInfo struct {
-	key        shared.IssuerKeyItf
+	key        IssuerKeyItf
 	issuertype string
 }
 
 // NewACMEIssuerInfo creates info for an ACME issuer
-func NewACMEIssuerInfo(key shared.IssuerKeyItf) IssuerInfo {
+func NewACMEIssuerInfo(key IssuerKeyItf) IssuerInfo {
 	return IssuerInfo{key: key, issuertype: IssuerTypeACME}
 }
 
 // NewCAIssuerInfo creates info for an CA issuer
-func NewCAIssuerInfo(key shared.IssuerKeyItf) IssuerInfo {
+func NewCAIssuerInfo(key IssuerKeyItf) IssuerInfo {
 	return IssuerInfo{key: key, issuertype: IssuerTypeCA}
 }
 
 // NewSelfSignedIssuerInfo creates info for a selfSigned issuer.
-func NewSelfSignedIssuerInfo(key shared.IssuerKeyItf) IssuerInfo {
+func NewSelfSignedIssuerInfo(key IssuerKeyItf) IssuerInfo {
 	return IssuerInfo{key: key, issuertype: IssuerTypeSelfSigned}
 }
 
 // Key returns the issuer key
-func (i *IssuerInfo) Key() shared.IssuerKeyItf {
+func (i *IssuerInfo) Key() IssuerKeyItf {
 	return i.key
 }
